Restructured profile page, extracted to separate views
