+++ Caffeine/Caffeine/ViewControllers/ProfileCell.swift
@@ -13,11 +13,10 @@ class ProfileCell: UICollectionViewCell {
     override init(frame: CGRect) {
         super.init(frame: frame)
     }
-    
-    func addContent(content: UIView){
+
+    func addContent(content: UIView) {
         self.addSubview(content)
         content.translatesAutoresizingMaskIntoConstraints = false
-        
         self.setContentConstraints(view: content)
     }
 
@@ -34,10 +33,6 @@ extension ProfileCell {
             view.centerYAnchor.constraint(equalTo: self.centerYAnchor),
             view.centerXAnchor.constraint(equalTo: self.centerXAnchor),
             view.heightAnchor.constraint(equalToConstant: self.contentView.frame.height)
-//            view.widthAnchor.constraint(equalToConstant: self.frame.width),
-//            view.centerYAnchor.constraint(equalTo: self.centerYAnchor),
-//            view.centerXAnchor.constraint(equalTo: self.centerXAnchor),
-//            view.heightAnchor.constraint(equalToConstant: self.frame.height)
             ])
     }
 }
